Support tables join between different group by pulling up to CN (merge request !66)
The original check relied too much on global variables and guc values, and mistakenly prevented the unique path from pulling to CN for calculation. We deal with it by removing guc and a global variable, and guard it at a more proper position. Same guard was added in the "INSERT INTO SELECT FROM" case.